+++ packages/_helpers/index.ts
@@ -1,2 +1,3 @@
 export { Clickable } from './clickable';
 export { DeadToggle } from './dead-toggle';
+export { Affix } from './affix';
