+++ packages/breadcrumbs/src/component.tsx
@@ -1,26 +1,26 @@
 import * as React from 'react';
 import { classNames } from '@chbphone55/classnames';
-import { interleave } from './utils';
 import type { BreadcrumbsProps } from './props';
 
-const setup = (props) => {
-  const { children, ariaLabel, className, ...rest } = props;
-  return {
-    attrs: {
-      'aria-label': ariaLabel || 'Her er du',
-      ...rest,
-    },
-    children: interleave(React.Children.toArray(children)),
-    classes: classNames('flex space-x-8', className),
-  };
-};
-
 export const Breadcrumbs = (props: BreadcrumbsProps) => {
-  const { classes, children, attrs } = setup(props);
+  const { children, className, ...rest } = props;
+  const ariaLabel = props['aria-label'] || 'Her er du';
+
   return (
-    <nav className={classes} {...attrs}>
-      <h2 className="u-screen-reader-only">{attrs['aria-label']}</h2>
-      {children}
+    <nav
+      className={classNames('flex space-x-8', props.className)}
+      aria-label={ariaLabel}
+      {...rest}
+    >
+      <h2 className="u-screen-reader-only">{ariaLabel}</h2>
+      {children?.map((crumb, i) => [
+        crumb,
+        i !== children.length - 1 && (
+          <span key={i} aria-hidden className="select-none">
+            /
+          </span>
+        ),
+      ])}
     </nav>
   );
 };
