+++ packages/combobox/docs/Combobox.mdx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Combobox } from '../src';
-import { Affix } from '../../textfield/src/affix';
+import { Affix } from '../../_helpers';
 
 # Combobox
 
@@ -214,6 +214,6 @@ export type ComboboxOption = {
 
 ## Affix Props
 
-```props packages/textfield/src/affix.tsx
+```props packages/_helpers/affix.tsx
 
 ```
