+++ packages/textfield/docs/TextField.mdx
@@ -1,4 +1,5 @@
-import { TextField, Affix } from '../src';
+import { TextField } from '../src';
+import { Affix } from '../../_helpers';
 
 # TextField
 
@@ -168,6 +169,6 @@ for more information.
 
 ## Affix Props
 
-```props packages/textfield/src/affix.tsx
+```props packages/_helpers/affix.tsx
 
 ```
