+++ packages/textfield/src/index.ts
@@ -1,3 +1,2 @@
 export { TextField } from './component';
-export { Affix } from './affix';
 export type { TextFieldProps } from './props';
